dsp: msm_audio_ion: Configure DMA API if IOMMU is enabled
When IOMMU assisted allocations are enabled, we need to get a
handle to an IOMMU master (specified in DT): this is being
actually done by calling arm_iommu_attach_device and that's ok.

The driver is not accounting for clean IOMMU drivers: it does
expect the IOMMU driver to auto-register its masters at probe
time which is not a common practice, infact the only driver
doing this is arm-smmu.
We need to use the msm_audio_ion driver with a new, clean and
Linux API adhering IOMMU driver (msm_iommu-v1) that is adding
the masters through the of_xlate IOMMU API callback: this
requires setting up the DMA-API mappings properly, which will
also configure the IOMMU through the appropriate standard API
calls.

Signed-off-by: Angelo G. Del Regno <kholk11@gmail.com>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>